데이터 타입 수정

+++ CalculatorProject/src/calculator/Lv2/DivideOperation.java
@@ -1,7 +1,7 @@
 package calculator.Lv2;
 
 public class DivideOperation {
-    public double operate(int firstNumber, int secondNumber){
+    public int operate(int firstNumber, int secondNumber){
         return firstNumber / secondNumber;
     }
 }
